+++ src/features/product-list/api/index.ts
@@ -1,0 +1,2 @@
+export { getOngoingProductList } from "./getOngoingProductList";
+export { getEnrollProductList } from "./getEnrollProductList";
