+++ src/features/product-list/ui/ProductListTabs.tsx
@@ -5,7 +5,7 @@ interface ProductListTabsProps {
   setActiveTab: (tab: string) => void;
 }
 
-const ProductListTabs = ({ activeTab, setActiveTab }: ProductListTabsProps) => {
+export const ProductListTabs = ({ activeTab, setActiveTab }: ProductListTabsProps) => {
   return (
     <div className="flex justify-center w-full mt-3">
       <div
@@ -33,5 +33,3 @@ const ProductListTabs = ({ activeTab, setActiveTab }: ProductListTabsProps) => {
     </div>
   );
 };
-
-export default ProductListTabs;
