+++ src/features/user/ui/index.ts
@@ -6,3 +6,7 @@ export { UserOrderList } from "./UserOrderList";
 export { UserOrderTab } from "./UserOrderTab";
 export { UserPreOrderTab } from "./UserPreOrderTab";
 export { UserProfile } from "./UserProfile";
+export { OrderHistoryProduct } from "./OrderHistoryProduct";
+export { OrderListTab } from "./OrderListTab";
+export { OrderLostProduct } from "./OrderLostProduct";
+export { OrderWonProduct } from "./OrderWonProduct";
